requirements.txt
streamlit>=1.28.0
ollama>=0.1.7
psycopg2-binary>=2.9.7
langchain>=0.1.0
langchain-text-splitters>=0.0.1
pypdf2>=3.0.1
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.3